Renaming MarathonSpecUtil to ServiceSpecUtil
