+++ src/js/components/modals/ServiceSuspendModal.js
@@ -3,11 +3,11 @@ import {Confirm} from 'reactjs-components';
 import React from 'react';
 /* eslint-enable no-unused-vars */
 
-import MarathonSpecUtil from '../../utils/MarathonSpecUtil';
 import MarathonStore from '../../stores/MarathonStore';
 import Pod from '../../structs/Pod';
 import ServiceTree from '../../structs/ServiceTree';
 import ServiceActionModal from './ServiceActionModal';
+import ServiceSpecUtil from '../../utils/ServiceSpecUtil';
 
 class ServiceSuspendModal extends ServiceActionModal {
   constructor() {
@@ -46,7 +46,7 @@ class ServiceSuspendModal extends ServiceActionModal {
       MarathonStore.editGroup({id: serviceID, scaleBy: 0}, forceUpdate);
     } else if (isPod) {
       MarathonStore.editService(service,
-        MarathonSpecUtil.Pods.setFixedScaling(service.getSpec(), 0),
+        ServiceSpecUtil.setFixedScaling(service.getSpec(), 0),
         forceUpdate );
     } else {
       MarathonStore.editService(service, {instances: 0}, forceUpdate);
